Remove Helper to merge Template
This helper is targeted to compatible Sprockets 2 with 3.
However current `ember-handlebars-template` supports only Sprockets >= 3.3.
So these files has not need to be separated.
