--- lib/ember/handlebars/helper.rb
@@ -1,69 +1,0 @@
-module Ember
-  module Handlebars
-    module Helper
-      def self.included(base)
-        base.extend ClassMethods
-      end
-
-      module ClassMethods
-        def setup_ember_template_compiler(path)
-          Barber::Ember::Precompiler.ember_template_compiler_path = path
-        end
-      end
-
-      private
-
-      def handlebars?(filename)
-        filename.to_s =~ /\.raw\.(handlebars|hjs|hbs)/
-      end
-
-      def mustache_to_handlebars(filename, template)
-        if filename =~ /\.mustache\.(handlebars|hjs|hbs)/
-          template.gsub(/\{\{(\w[^\}]+)\}\}/){ |x| "{{unbound #{$1}}}" }
-        else
-          template
-        end
-      end
-
-      def amd_template_target(namespace, module_name)
-        [namespace, module_name].compact.join('/')
-      end
-
-      def global_template_target(namespace, module_name, config)
-        "#{namespace}[#{template_path(module_name, config).inspect}]"
-      end
-
-      def template_path(path, config)
-        root = config.templates_root
-
-        unless root.empty?
-          Array(root).each.each do |r|
-            path = path.sub(/#{Regexp.quote(r)}\//, '')
-          end
-        end
-
-        path.split('/').join(config.templates_path_separator)
-      end
-
-      def compile_handlebars(string)
-        "Handlebars.compile(#{indent(string).inspect});"
-      end
-
-      def precompile_handlebars(string)
-        "Handlebars.template(#{Barber::Precompiler.compile(string)});"
-      end
-
-      def compile_ember_handlebars(string, ember_template = 'Handlebars', options = nil)
-        "Ember.#{ember_template}.compile(#{indent(string).inspect}, #{options.to_json});"
-      end
-
-      def precompile_ember_handlebars(string, ember_template = 'Handlebars', options = nil)
-        "Ember.#{ember_template}.template(#{Barber::Ember::Precompiler.compile(string, options)});"
-      end
-
-      def indent(string)
-        string.gsub(/$(.)/m, "\\1  ").strip
-      end
-    end
-  end
-end
