+++ lib/ember/handlebars/template.rb
@@ -5,10 +5,8 @@ module Ember
   module Handlebars
     autoload :VERSION, 'ember/handlebars/version'
     autoload :Config, 'ember/handlebars/config'
-    autoload :Helper, 'ember/handlebars/helper'
 
     class Template
-      include Helper
 
       class << self
         def configure
@@ -28,6 +26,10 @@ module Ember
 
           env.register_mime_type 'text/x-ember-handlebars', extensions: with_js_extension(%w(.hbs .hjs .handlebars))
           env.register_transformer 'text/x-ember-handlebars', 'application/javascript', self
+        end
+
+        def setup_ember_template_compiler(path)
+          Barber::Ember::Precompiler.ember_template_compiler_path = path
         end
 
         def instance
@@ -110,6 +112,38 @@ module Ember
 
       private
 
+      def handlebars?(filename)
+        filename.to_s =~ /\.raw\.(handlebars|hjs|hbs)/
+      end
+
+      def mustache_to_handlebars(filename, template)
+        if filename =~ /\.mustache\.(handlebars|hjs|hbs)/
+          template.gsub(/\{\{(\w[^\}]+)\}\}/){ |x| "{{unbound #{$1}}}" }
+        else
+          template
+        end
+      end
+
+      def amd_template_target(namespace, module_name)
+        [namespace, module_name].compact.join('/')
+      end
+
+      def global_template_target(namespace, module_name, config)
+        "#{namespace}[#{template_path(module_name, config).inspect}]"
+      end
+
+      def template_path(path, config)
+        root = config.templates_root
+
+        unless root.empty?
+          Array(root).each.each do |r|
+            path = path.sub(/#{Regexp.quote(r)}\//, '')
+          end
+        end
+
+        path.split('/').join(config.templates_path_separator)
+      end
+
       def precompile_handlebars(template, input)
         dependencies = [
           Barber::Precompiler.compiler_version,
@@ -117,7 +151,7 @@ module Ember
         ]
 
         input[:cache].fetch(_cache_key + dependencies) do
-          super(template)
+          "Handlebars.template(#{Barber::Precompiler.compile(template)});"
         end
       end
 
@@ -130,8 +164,20 @@ module Ember
         ]
 
         input[:cache].fetch(_cache_key + dependencies) do
-          super(template, ember_template, options)
+          "Ember.#{ember_template}.template(#{Barber::Ember::Precompiler.compile(template, options)});"
         end
+      end
+
+      def compile_handlebars(string)
+        "Handlebars.compile(#{indent(string).inspect});"
+      end
+
+      def compile_ember_handlebars(string, ember_template, options = nil)
+        "Ember.#{ember_template}.compile(#{indent(string).inspect}, #{options.to_json});"
+      end
+
+      def indent(string)
+        string.gsub(/$(.)/m, "\\1  ").strip
       end
 
       def _cache_key
